Rewrite `View::Email` to use `Email::Stuffer` for email sending
`Catalyst::View::Email` and `Catalyst::View::Email::Template` are old,
buggy and the maintainer announced, that he won't work on the project
anymore. With this commit we have a new `Coocook::View::Email` which
uses the simple API of `Email::Stuffer`, which also uses
`Email::Sender::Simple` under the hood, to send emails.
Because `Email::Stuffer` uses `Email::Sender::Emails` our current test
and debug configurations don't need to be changed and work like before,
so that no real emails are sent in debug or test mode.