Remove FK from {Dish,Recipe}Ingredient on ArticleUnit
ArticleUnit should not be required anymore for an Ingredient, because
there are cases where for example a user adds a newly created Unit to
an Ingredient with an existing Article. In this case the Unit should not
get implicitly connected with the Article, because this would pollute
the Units of the Article and is implicit behaviour that the user might
not expect or want. To protect the user from this it should be possible
to add a Unit just to an Ingredient, e.g. as an Exception, but not to
the article. To realize this technically this the FK from the two Ingredient
results on ArticleUnit must be removed.

The purpose of the ArticleUnit table is now defined as: 'The Units with an
ArticleUnit row are the suggested Units for an Article.'