Optimize migration: generate "position" while population temp table
The order if dishes in table "dishes" had no explicit order before.
Dishes were shown in the order as delivered by PostgreSQL, in
many cases the order of creation of dishes.

In order to preserve this order as far as possible we move the
generation of values for the new "position" column to the
population of the temporary table. So the internal order of
rows in the temporary table has no effect anymore.